feature(adroll): address comments: refactor
